refactor: move all errors imports a single file
