+++ src/presentation/controllers/signup.spec.ts
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { SignUpController } from './signup'
-import { MissingParamError } from '../errors/missing-param-error'
-import { InvalidParamError } from '../errors/invalid-param-error'
 import { EmailValidator } from '../protocols/email-validator'
-import { ServerError } from '../errors/server-error'
+import { MissingParamError, ServerError, InvalidParamError } from '../errors'
+
 
 interface SutTypes {
   sut: SignUpController
